refactor: Extract info icon tooltip
