+++ app/configurator/components/chart-type-selector.tsx
@@ -1,12 +1,5 @@
-import { Trans, t } from "@lingui/macro";
-import {
-  Box,
-  BoxProps,
-  Divider,
-  Theme,
-  Tooltip,
-  Typography,
-} from "@mui/material";
+import { t, Trans } from "@lingui/macro";
+import { Box, BoxProps, Divider, Theme, Typography } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import React, { useCallback } from "react";
 
@@ -27,8 +20,9 @@ import { IconButton } from "@/configurator/components/icon-button";
 import { useChartType } from "@/configurator/config-form";
 import { ConfiguratorStateWithChartConfigs } from "@/configurator/configurator-state";
 import { useDataCubesComponentsQuery } from "@/graphql/hooks";
-import { Icon } from "@/icons";
 import { useLocale } from "@/locales/use-locale";
+
+import { InfoIconTooltip } from "./info-icon-tooltip";
 
 type ChartTypeSelectorProps = {
   state: Exclude<ConfiguratorStateWithChartConfigs, ConfiguratorStatePublished>;
@@ -214,37 +208,9 @@ const ChartTypeSelectorMenu = (props: ChartTypeSelectorMenuProps) => {
   );
 };
 
-const useWarnIconStyles = makeStyles<Theme>((theme) => ({
+export const useWarnIconStyles = makeStyles<Theme>((theme) => ({
   icon: {
     color: theme.palette.primary.main,
     pointerEvents: "auto",
   },
 }));
-
-type InfoIconTooltipProps = {
-  title: NonNullable<React.ReactNode>;
-};
-
-const InfoIconTooltip = (props: InfoIconTooltipProps) => {
-  const { title } = props;
-  const iconStyles = useWarnIconStyles();
-
-  return (
-    <Tooltip
-      arrow
-      placement="top"
-      title={
-        <Typography variant="caption" color="secondary">
-          {title}
-        </Typography>
-      }
-      componentsProps={{
-        tooltip: { sx: { width: 180, px: 2, py: 1, lineHeight: "18px" } },
-      }}
-    >
-      <Typography>
-        <Icon name="infoOutline" size={16} className={iconStyles.icon} />
-      </Typography>
-    </Tooltip>
-  );
-};
